Fix UploadStreamToBlockBlob not reporting incoming stream errors
Before this change, when using UploadStreamToBlockBlob, if the input
stream returned an error UploadStreamToBlockBlob would not return an
error and create a blob.

This is incorrect - when there is an error in the input stream
UploadStreamToBlockBlob should return an error and the blob should not
be created.

This fix makes sure the error is reported correctly and the blob is
not created in the input stream.

Fixes #52