Reduces construction of chunk ids for the commit list, moves azblob_test to azblob
azblob_test should be azblob test, allowing access to private types and removing the need for azblob. as a prefix.  Could find no reason to make as a seperate package in a non-go standard way.

This packages' derivation from the blobstore standard of a new UUID per chunk has merits as discussed with Adele(faster, less memory, possible upload resumes, etc...).  So it was decided to keep it.

However, wanted to make this easier for us to autoincrement ids and provide the list of IDs to commit instead of recreating them at the end at the cost of CPU we didn't have to spend (we were going to spend the memory anyways).  So provided a better way to get the IDs.

This change required changes to the tests.  Most tests use a single blockID.  For those tests we now use a var created on init() that is a UUID+ math.MaxUint32.  That allows us to test the maximum value.

For others, we now use our id type.  This changed one test that was trying to test order, which wasn't necessary.

All tests are passing.