Update go.mod via go tidy command
The adal change was because one of the zt tests uses it.  It always should have been there and won't cause any change in functionality.

errors gets added as an indirecty from check.v1, which it should have always had. This is because check doesn't have a go.mod file.

Adds a minimum version of Go compiler 1.13