[chunkwriting] Handle data races in chunk IDs (#174)
During test execution `TestCopyFromReader` hit a data race which
resulted in the test failing due to checksum mismatch.

Since the `id` struct's methods are not thread-safe there is a race
condition where two upload threads request an ID at the same moment and
both could potentially end up with the same ID.

The simplest solution is to move the ID assignment to the thread that
populates the chunks to the channel. This ensures that the chunk IDs are
assigned sequentially and removes the race condition.